End of day commit
- Starting to introduce a trader service

+++ src/strategy.domain/WeightedMovingAverageCalculator.cs
@@ -6,9 +6,9 @@ namespace strategy.domain
 {
     public class WeightedMovingAverageCalculator
     {
-        public decimal Calculate(IEnumerable<TickerPrice> prices, int weight)
+        public double Calculate(IEnumerable<TickerPrice> prices, int weight)
         {
-            decimal wma = 0;
+            double wma = 0;
             var dataPoints = prices
                 .OrderByDescending(x => x.DateTime)
                 .Take(weight)
@@ -17,7 +17,7 @@ namespace strategy.domain
 
             for (int i = 0; i < dataPoints.Length; i++)
             {
-                decimal w = (i + 1) / (decimal)weight;
+                double w = (i + 1) / (double)weight;
                 var ma = dataPoints[i].Price * w;
                 wma += ma;
             }
